Refactor / rework / documentation
Refactoring the code for this project into a more modular way.

Additional, I have reworked the functionality to of this project to use strings rather than integers, also the program now checks the user's input and rerun the program when the user picks an invalid option.

Lastly, I re did the documentation of the project to be better.